conveyor belt with dynamic center of mass so boxes do not roll
If center of mass is in the nucleus of the boxes, they just roll down the conveyor belt (since they're being commanded to move by some value of x, and that's the most physics-based way to move them). When added to the list of items to convey, the conveyor tells the box instance to move its center of mass (to a dumb number). When the box is no longer touching the conveyor, the conveyor messages the box to put its CoM back to its default value. Need computations on current rotation to make up a better modified CoM.